Closed it off nicely
As apis are largely withdrawn now, a redirect to the about page has been put on. Along with an explination of this.

Should it be needed the API's can be brought back fairly simply and the rest of the system sorted out,as there was still a smattering of errors around.